pyproject.toml
[project]
name = "gpt"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.135.1",
    "numpy>=2.4.2",
    "onnxruntime>=1.24.2",
    "requests>=2.32.5",
    "uvicorn>=0.41.0",
]

[project.optional-dependencies]
train = [
    "hydra-core>=1.3.2",
    "lightning>=2.6.1",
    "omegaconf>=2.3.0",
    "onnx>=1.20.1",
    "onnxscript>=0.6.2",
    "pytorch-lightning>=2.6.1",
    "torch>=2.10.0",
]